fixed bugs (int to size_t)

+++ sort.h
@@ -27,6 +27,6 @@ void bubble_sort(int *array, size_t size);
 void insertion_sort_list(listint_t **list);
 void swap(listint_t *a, listint_t *b, listint_t **head);
 void selection_sort(int *array, size_t size);
-int sort_min(int *array, size_t start, size_t size);
+size_t sort_min(int *array, size_t start, size_t size);
 
 #endif
